Added new method of copying files between source, tmp and target. Added new `regolith watch` command.

+++ regolith/main_functions.go
@@ -4,6 +4,7 @@ import (
 	"encoding/json"
 	"io/ioutil"
 	"os"
+	"path/filepath"
 )
 
 // Install handles the "regolith install" command. It installs specific filters
@@ -151,26 +152,77 @@ func UpdateAll(debug bool) error {
 	return nil
 }
 
+// runOrWatch handles both 'regolith run' and 'regolith watch' commands based
+// on the 'watch' parameter. It runs/watches the profile named after
+// 'profileName' parameter. The 'debug' argument determines if the debug
+// messages should be printed or not.
+func runOrWatch(profileName string, debug, watch bool) error {
+	InitLogging(debug)
+	if profileName == "" {
+		profileName = "dev"
+	}
+	// Load the Config and the profile
+	configJson, err := LoadConfigAsMap()
+	if err != nil {
+		return WrapError(err, "Could not load \"config.json\".")
+	}
+	config, err := ConfigFromObject(configJson)
+	if err != nil {
+		return WrapError(err, "Could not load \"config.json\".")
+	}
+	profile, ok := config.Profiles[profileName]
+	if !ok {
+		return WrappedErrorf(
+			"Profile %q does not exist in the configuration.", profileName)
+	}
+	// Check the filters of the profile
+	err = CheckProfileImpl(profile, profileName, *config, nil)
+	if err != nil {
+		return err
+	}
+	path, _ := filepath.Abs(".")
+	context := RunContext{
+		AbsoluteLocation: path,
+		Config:           config,
+		Parent:           nil,
+		Profile:          profileName,
+	}
+	if watch { // Loop until program termination (CTRL+C)
+		context.StartWatchingSrouceFiles()
+		for {
+			err = RunProfile(context)
+			if err != nil {
+				Logger.Errorf(
+					"Failed to run profile %q: %s",
+					profileName, PassError(err).Error())
+			} else {
+				Logger.Infof("Successfully ran the %q profile.", profileName)
+			}
+			Logger.Info("Press Ctrl+C to stop watching.")
+			context.AwaitInterruption()
+			Logger.Warn("Restarting...")
+		}
+		// return nil // Unreachable code
+	}
+	err = RunProfile(context)
+	if err != nil {
+		return WrapErrorf(err, "Failed to run profile %q", profileName)
+	}
+	Logger.Infof("Successfully ran the %q profile.", profileName)
+	return nil
+}
+
 // Run handles the "regolith run" command. It runs selected profile and exports
 // created resource pack and behvaiour pack to the target destination.
-//
-// The "profile" parameter is the name of the profile to run. If the profile
-// is an empty string, the "dev" profile will be used.
-//
-// The "debug" parameter is a boolean that determines if the debug messages
-// should be printed.
-func Run(profile string, debug bool) error {
-	InitLogging(debug)
-	if profile == "" {
-		profile = "dev"
-	}
-	Logger.Infof("Running %q profile...", profile)
-	err := RunProfile(profile)
-	if err != nil {
-		return WrapErrorf(err, "Failed to run profile %q", profile)
-	}
-	Logger.Infof("Successfully ran the %q profile.", profile)
-	return nil
+func Run(profileName string, debug bool) error {
+	return runOrWatch(profileName, debug, false)
+}
+
+// Watch handles the "regolith watch" command. It watches the project
+// directories and it runs selected profile and exports created resource pack
+// and behvaiour pack to the target destination when the project changes.
+func Watch(profileName string, debug bool) error {
+	return runOrWatch(profileName, debug, true)
 }
 
 // Init handles the "regolith init" command. It initializes a new Regolith
@@ -243,16 +295,23 @@ func Init(debug bool) error {
 //
 // The "debug" parameter is a boolean that determines if the debug messages
 // should be printed.
-func Clean(debug bool) error {
+func Clean(debug bool, cachedStatesOnly bool) error {
 	InitLogging(debug)
 	Logger.Infof("Cleaning cache...")
-	err := os.RemoveAll(".regolith")
-	if err != nil {
-		return WrapError(err, "failed to remove .regolith folder")
-	}
-	err = os.Mkdir(".regolith", 0666)
-	if err != nil {
-		return WrapError(err, "failed to recreate .regolith folder")
+	if cachedStatesOnly {
+		err := ClearCachedStates()
+		if err != nil {
+			return WrapError(err, "Failed to remove cached path states.")
+		}
+	} else {
+		err := os.RemoveAll(".regolith")
+		if err != nil {
+			return WrapError(err, "failed to remove .regolith folder")
+		}
+		err = os.Mkdir(".regolith", 0666)
+		if err != nil {
+			return WrapError(err, "failed to recreate .regolith folder")
+		}
 	}
 	Logger.Infof("Cache cleaned.")
 	return nil
